1) Change definition of a "Close Score" to be the losing team having a 10% or greater chance of winning.
Through looking at teams, I realized what I want is games that are still "in play", as opposed to games that are especially close.

For example, there's a game with only a 12% chance to win, where the winning team has no players left to play, and the losing team needs 17.63 to tie, with David Njoku left. That feels like it's still a competitive game worth following.

2) Modified the calculation of whether or not a score is close for readibility. Now it does the min of the two winProbability variables and checks to see if it's greater or equal to 10%. No mental math is required anymore. just plug in the threshold the losing team needs to be considered a "close game".